Changed discount amount to integer instead of BigDecimal

+++ src/main/java/com/app/entities/DiscountCode.java
@@ -26,7 +26,7 @@ public class DiscountCode {
     private Long id;
 
     private String code;
-    private BigDecimal discountAmount;
+    private Integer discountAmount;
 
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
     private LocalDate expiryDate;
